requirements.txt
python-dotenv
openai-whisper
requests
pandas
whisper
numpy
matplotlib
python-Levenshtein
jiwer
